+++ seshat-node/src/utils.rs
@@ -105,6 +105,10 @@ pub(crate) fn parse_search_object(
         config.order_by_recency(v.value(cx));
     }
 
+    if let Some(v) = argument.get_opt::<JsBoolean, _, _>(&mut *cx, "query_syntax")? {
+        config.query_syntax(v.value(cx));
+    }
+
     if let Some(r) = argument.get_opt::<JsString, _, _>(&mut *cx, "room_id")? {
         config.for_room(&r.value(cx));
     }